# Conflicts:
#	data/battle_anim_scripts.s
#	data/battle_scripts_1.s
#	include/battle_scripts.h
#	include/battle_util.h
#	src/battle_script_commands.c
#	src/battle_util.c

+++ include/battle_util.h
@@ -192,7 +192,7 @@ void ClearIllusionMon(u32 battler);
 bool32 SetIllusionMon(struct Pokemon *mon, u32 battler);
 bool8 ShouldGetStatBadgeBoost(u16 flagId, u8 battler);
 u8 GetBattleMoveSplit(u32 moveId);
-bool32 CanFling(u8 battler);
+bool32 CanFling(u32 battler);
 bool32 IsTelekinesisBannedSpecies(u16 species);
 bool32 IsHealBlockPreventingMove(u32 battler, u32 move);
 bool32 HasEnoughHpToEatBerry(u32 battler, u32 hpFraction, u32 itemId);
